Calculator/Calculator.Application/CalculatorService.cs
using Calculator.Application.Interfaces;

namespace Calculator.Application;

public class CalculatorService: ICalculatorService
{
    public double Add(double a, double b)
    {
        return a + b;
    }

    public double Subtract(double a, double b)
    {
        return a - b;
    }

    public double Multiply(double a, double b)
    {
        return a * b;
    }

    public double Divide(double a, double b)
    {
        if (b == 0)
            throw new DivideByZeroException();

        return a / b;
    }
}